+++ changelog.md
@@ -1,6 +1,6 @@
 # 1.7.5
 
-- support `title` fields with spaces [src](https://github.com/xddq/schema2typebox/pull/53)
+- support `title` fields containing characters '- .\_' [src](https://github.com/xddq/schema2typebox/pull/53)
 
 # 1.7.5